mnt/user-data/outputs/ultravox-mcp-unofficial/requirements.txt
# Ultravox MCP - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies (required)
python-dotenv>=1.0.0           # Environment variable management
httpx>=0.24.0                  # Modern HTTP client
requests>=2.31.0               # HTTP library (fallback)

# Optional but recommended
fastmcp>=0.0.1                 # FastMCP implementation (auto-installs)
pydantic>=2.0.0                # Data validation
typing-extensions>=4.8.0       # Type hints

# Development & Testing (optional)
pytest>=7.4.0                  # Testing framework
pytest-asyncio>=0.21.0         # Async test support
pytest-cov>=4.1.0              # Coverage reporting
black>=23.0.0                  # Code formatter
flake8>=6.0.0                  # Linter
mypy>=1.5.0                    # Type checker

# Documentation (optional)
sphinx>=7.0.0                  # Documentation generator
sphinx-rtd-theme>=1.3.0        # ReadTheDocs theme
mkdocs>=1.5.0                  # Alternative docs
mkdocs-material>=9.0.0         # Material theme

# Logging (optional)
python-json-logger>=2.0.0      # JSON logging
colorama>=0.4.6                # Colored console output

# Note: FastMCP is optional - the server includes a stdlib fallback
# If FastMCP is not available, it will automatically use the stdlib version
